refactor(desktop): extract trackTerminalActivity to own file for clean test imports
