+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/TerminalPane/hooks/useTerminalActivityTracking/useTerminalActivityTracking.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, mock, test } from "bun:test";
-import { trackTerminalActivity } from "./useTerminalActivityTracking";
+import { trackTerminalActivity } from "./trackTerminalActivity";
 
 describe("trackTerminalActivity", () => {
 	let updateLastActivityAt: ReturnType<typeof mock>;
